+++ tests/tokenizer_test.php
@@ -74,6 +74,14 @@ class tokenizer_test extends \advanced_testcase {
     protected static array $invalidpreparsecases;
 
     /**
+     * List of test cases for tokenizer::init_set_max_token_count
+     *
+     * key: path of a testable JSON file
+     * value: expected value for max_token_count
+     */
+    protected static array $maxtokencountcases;
+
+    /**
      * List of test cases for tokenizer::init_override_tokens
      *
      * key: path of a testable JSON file
@@ -134,6 +142,7 @@ class tokenizer_test extends \advanced_testcase {
      */
     public static function setUpBeforeClass(): void {
         self::setup_invalid_cases();
+        self::setup_max_token_count_cases();
         self::setup_override_tokens_cases();
         self::setup_merge_cases();
         self::setup_prepare_cases();
@@ -171,7 +180,7 @@ class tokenizer_test extends \advanced_testcase {
      */
     public function test_static_check() {
         try {
-            $dir = dirname(__FILE__) . '/../similarity/rules';
+            $dir = dirname(__FILE__) . '/../similarity/tokenizer_rules';
 
             $scanarr = scandir($dir);
             $filesarr = array_diff($scanarr, array('.', '..'));
@@ -193,8 +202,6 @@ class tokenizer_test extends \advanced_testcase {
         }
     }
 
-
-
     /**
      * Method to test tokenizer::init with invalid files
      */
@@ -209,6 +216,17 @@ class tokenizer_test extends \advanced_testcase {
             }
 
             $this->fail('An expection was expected');
+        }
+    }
+
+    /**
+     * Method to test tokenizer::set_max_token_count
+     */
+    public function test_set_max_token_count() {
+        foreach (self::$maxtokencountcases as $filename => $expectedmaxtokencount) {
+            $tokenizer = new tokenizer($filename);
+            $maxtokencount = testable_tokenizer::get_max_token_count_from($tokenizer);
+            $this->assertSame($expectedmaxtokencount, $maxtokencount);
         }
     }
 
@@ -419,93 +437,99 @@ class tokenizer_test extends \advanced_testcase {
                 'file ' . self::testpath()  . 'invalid/dump_test.json must exist'
             ),
             self::testpath() . 'invalid/general/not_good_suffix.json' => (
-                self::testpath() . 'invalid/general/not_good_suffix.json' . ' must have suffix _highlight_rules.json'
+                self::testpath() . 'invalid/general/not_good_suffix.json' . ' must have suffix _tokenizer_rules.json'
             ),
-            self::testpath() . 'invalid/general/empty_highlight_rules.json' => (
-                'file ' . self::testpath() . 'invalid/general/empty_highlight_rules.json' . ' is empty'
+            self::testpath() . 'invalid/general/empty_tokenizer_rules.json' => (
+                'file ' . self::testpath() . 'invalid/general/empty_tokenizer_rules.json' . ' is empty'
             ),
-            self::testpath() . 'invalid/general/undefined_option_highlight_rules.json' => (
+            self::testpath() . 'invalid/general/undefined_option_tokenizer_rules.json' => (
                 'invalid options: example'
             ),
-            self::testpath() . 'invalid/general/invalid_check_rules_highlight_rules.json' => (
+            self::testpath() . 'invalid/general/invalid_check_rules_tokenizer_rules.json' => (
                 '"check_rules" option must be a boolean'
             ),
-            self::testpath() . 'invalid/general/invalid_name_highlight_rules.json' => (
+            self::testpath() . 'invalid/general/invalid_max_token_count_tokenizer_rules.json' => (
+                '"max_token_count" option must be numeric'
+            ),
+            self::testpath() . 'invalid/general/invalid_number_max_token_count_tokenizer_rules.json' => (
+                '"max_token_count" option must be a positive integer'
+            ),
+            self::testpath() . 'invalid/general/invalid_name_tokenizer_rules.json' => (
                 '"name" option must be a string'
             ),
-            self::testpath() . 'invalid/general/invalid_extension_no_string_highlight_rules.json' => (
+            self::testpath() . 'invalid/general/invalid_extension_no_string_tokenizer_rules.json' => (
                 '"extension" option must be a string or an array of strings'
             ),
-            self::testpath() . 'invalid/general/invalid_extension_no_array_highlight_rules.json' => (
+            self::testpath() . 'invalid/general/invalid_extension_no_array_tokenizer_rules.json' => (
                 '"extension" option must be a string or an array of strings'
             ),
-            self::testpath() . 'invalid/general/invalid_extension_no_dot_highlight_rules.json' => (
+            self::testpath() . 'invalid/general/invalid_extension_no_dot_tokenizer_rules.json' => (
                 'extension c must start with .'
             ),
-            self::testpath() . 'invalid/general/invalid_inherit_rules_highlight_rules.json' => (
+            self::testpath() . 'invalid/general/invalid_inherit_rules_tokenizer_rules.json' => (
                 '"inherit_rules" option must be a string'
             ),
-            self::testpath() . 'invalid/states/invalid_data_states_highlight_rules.json' => (
+            self::testpath() . 'invalid/states/invalid_data_states_tokenizer_rules.json' => (
                 '"states" option must be an object'
             ),
-            self::testpath() . 'invalid/states/states_with_no_name_highlight_rules.json' => (
+            self::testpath() . 'invalid/states/states_with_no_name_tokenizer_rules.json' => (
                 'state 0 must have a name'
             ),
-            self::testpath() . 'invalid/states/state_not_object_highlight_rules.json' => (
+            self::testpath() . 'invalid/states/state_not_object_tokenizer_rules.json' => (
                 'state 0 must be an array'
             ),
-            self::testpath() . 'invalid/states/one_state_with_no_name_highlight_rules.json' => (
+            self::testpath() . 'invalid/states/one_state_with_no_name_tokenizer_rules.json' => (
                 'state 1 must have a name'
             ),
-            self::testpath() . 'invalid/rules/invalid_rule_highlight_rules.json' => (
+            self::testpath() . 'invalid/rules/invalid_rule_tokenizer_rules.json' => (
                 'rule 0 of state "state1" nº0 must be an object'
             ),
-            self::testpath() . 'invalid/rules/invalid_rule_option_value_highlight_rules.json' => (
+            self::testpath() . 'invalid/rules/invalid_rule_option_value_tokenizer_rules.json' => (
                 'invalid data type for token at rule 0 of state "state1" nº0'
             ),
-            self::testpath() . 'invalid/rules/undefined_rule_option_highlight_rules.json' => (
+            self::testpath() . 'invalid/rules/undefined_rule_option_tokenizer_rules.json' => (
                 'invalid option example at rule 0 of state "state1" nº0'
             ),
-            self::testpath() . 'invalid/rules/invalid_next_highlight_rules.json' => (
+            self::testpath() . 'invalid/rules/invalid_next_tokenizer_rules.json' => (
                 'invalid data type for next at rule 0 of state "state1" nº0'
             ),
-            self::testpath() . 'invalid/rules/regex_not_found_highlight_rules.json' => (
+            self::testpath() . 'invalid/rules/regex_not_found_tokenizer_rules.json' => (
                 'option token must be defined next to regex at rule 0 of state "state1" nº0'
             ),
-            self::testpath() . 'invalid/rules/token_not_found_highlight_rules.json' => (
+            self::testpath() . 'invalid/rules/token_not_found_tokenizer_rules.json' => (
                 'option regex must be defined next to token at rule 0 of state "state1" nº0'
             ),
-            self::testpath() . 'invalid/rules/invalid_token_value_highlight_rules.json' => (
+            self::testpath() . 'invalid/rules/invalid_token_value_tokenizer_rules.json' => (
                 'invalid token at rule 0 of state "start" nº0'
             ),
-            self::testpath() . 'invalid/rules/invalid_default_token_highlight_rules.json' => (
+            self::testpath() . 'invalid/rules/invalid_default_token_tokenizer_rules.json' => (
                 'invalid data type for default_token at rule 0 of state "start" nº0'
             ),
-            self::testpath() . 'invalid/rules/default_token_not_alone_highlight_rules.json' => (
+            self::testpath() . 'invalid/rules/default_token_not_alone_tokenizer_rules.json' => (
                 'option default_token must be alone at rule 0 of state "start" nº0'
             ),
-            self::testpath() . 'invalid/general/invalid_json_inheritance_highlight_rules.json' => (
-                'inherit JSON file ' . self::testpath() . 'invalid/general/dump_highlight_rules.json does not exist'
+            self::testpath() . 'invalid/general/invalid_json_inheritance_tokenizer_rules.json' => (
+                'inherit JSON file ' . self::testpath() . 'invalid/general/dump_tokenizer_rules.json does not exist'
             ),
-            self::testpath() . 'invalid/general/invalid_override_tokens_highlight_rules.json' => (
+            self::testpath() . 'invalid/general/invalid_override_tokens_tokenizer_rules.json' => (
                 '"override_tokens" option must be an object'
             ),
-            self::testpath() . 'invalid/general/invalid_token_at_override_tokens_highlight_rules.json' => (
+            self::testpath() . 'invalid/general/invalid_token_at_override_tokens_tokenizer_rules.json' => (
                 'this_is_not_a_good_token_name does not exist'
             ),
-            self::testpath() . 'invalid/general/vpl_type_not_overrided_highlight_rules.json' => (
+            self::testpath() . 'invalid/general/vpl_type_not_overrided_tokenizer_rules.json' => (
                 'vpl_literal could not be overrided'
             )
         );
 
         self::$invalidpreparsecases = array(
-            self::testpath() . 'invalid/general/invalid_file_at_preparse_highlight_rules.json' => (
+            self::testpath() . 'invalid/general/invalid_file_at_preparse_tokenizer_rules.json' => (
                 [
                     'input' => 'dump_test.js',
                     'output' => 'file dump_test.js does not exist',
                 ]
             ),
-            self::testpath() . 'invalid/general/invalid_ext_at_preparse_highlight_rules.json' => (
+            self::testpath() . 'invalid/general/invalid_ext_at_preparse_tokenizer_rules.json' => (
                 [
                     'input' => self::testpath() . 'invalid/general/test_ext.java',
                     'output' => self::testpath() . 'invalid/general/test_ext.java must end with one of the extensions .c,.h',
@@ -514,30 +538,44 @@ class tokenizer_test extends \advanced_testcase {
         );
     }
 
+    private static function setup_max_token_count_cases(): void {
+        self::$maxtokencountcases = array(
+            self::testpath() . 'valid/max_token_count/max_token_count_zero_tokenizer_rules.json' => 0,
+            self::testpath() . 'valid/max_token_count/max_token_count_ten_tokenizer_rules.json' => 10
+        );
+    }
+
     private static function setup_override_tokens_cases(): void {
         self::$overridetokenscases = array(
-            self::testpath() . 'valid/override_tokens/empty_override_tokens_highlight_rules.json' => [ ],
-            self::testpath() . 'valid/override_tokens/one_override_token_highlight_rules.json' => (
+            self::testpath() . 'valid/override_tokens/empty_override_tokens_tokenizer_rules.json' => [ ],
+            self::testpath() . 'valid/override_tokens/one_override_token_tokenizer_rules.json' => (
                 [
                     'comment' => token_type::LITERAL
                 ]
             ),
-            self::testpath() . 'valid/override_tokens/two_override_token_highlight_rules.json' => (
+            self::testpath() . 'valid/override_tokens/two_override_token_tokenizer_rules.json' => (
                 [
                     'comment' => token_type::LITERAL,
                     'comment.line' => token_type::LITERAL
                 ]
             ),
-            self::testpath() . 'valid/override_tokens/two_complex_override_token_highlight_rules.json' => (
+            self::testpath() . 'valid/override_tokens/two_complex_override_token_tokenizer_rules.json' => (
                 [
                     'string.start' => token_type::LITERAL,
                     'string.end' => token_type::LITERAL
                 ]
             ),
-            self::testpath() . 'valid/override_tokens/complex_override_token_highlight_rules.json' => (
+            self::testpath() . 'valid/override_tokens/complex_override_token_tokenizer_rules.json' => (
                 [
                     'string.start' => token_type::LITERAL,
                     'string.end' => token_type::LITERAL
+                ]
+            ),
+            self::testpath() . 'valid/override_tokens/inheritance_override_tokens_tokenizer_rules.json' => (
+                [
+                    'string.start' => token_type::LITERAL,
+                    'string.end' => token_type::LITERAL,
+                    'comment' => token_type::LITERAL
                 ]
             )
         );
@@ -545,27 +583,27 @@ class tokenizer_test extends \advanced_testcase {
 
     private static function setup_merge_cases(): void {
         self::$mergetestcases = array(
-            self::testpath() . 'valid/merge/merge_one_to_one_state_highlight_rules.json' => (
+            self::testpath() . 'valid/merge/merge_one_to_one_state_tokenizer_rules.json' => (
                 [
                     "start" => [ 0 => (object)[ "token" => "comment", "regex" => "\\/\\/", "next" => "text-state" ] ],
                     "text-state" => [ 0 => (object)[ "token" => "text", "regex" => ".*" ] ]
                 ]
             ),
-            self::testpath() . 'valid/merge/merge_one_to_two_states_highlight_rules.json' => (
-                [
-                    "start" => [ 0 => (object)[ "token" => "comment", "regex" => "\\/\\/", "next" => "text-state" ] ],
-                    "eol" => [ 0 => (object)[ "token" => "eol", "regex" => "\n" ] ],
-                    "text-state" => [ 0 => (object)[ "token" => "text", "regex" => ".*" ] ]
-                ]
-            ),
-            self::testpath() . 'valid/merge/merge_two_to_one_states_highlight_rules.json' => (
+            self::testpath() . 'valid/merge/merge_one_to_two_states_tokenizer_rules.json' => (
                 [
                     "start" => [ 0 => (object)[ "token" => "comment", "regex" => "\\/\\/", "next" => "text-state" ] ],
                     "eol" => [ 0 => (object)[ "token" => "eol", "regex" => "\n" ] ],
                     "text-state" => [ 0 => (object)[ "token" => "text", "regex" => ".*" ] ]
                 ]
             ),
-            self::testpath() . 'valid/merge/merge_with_same_states_highlight_rules.json' => (
+            self::testpath() . 'valid/merge/merge_two_to_one_states_tokenizer_rules.json' => (
+                [
+                    "start" => [ 0 => (object)[ "token" => "comment", "regex" => "\\/\\/", "next" => "text-state" ] ],
+                    "eol" => [ 0 => (object)[ "token" => "eol", "regex" => "\n" ] ],
+                    "text-state" => [ 0 => (object)[ "token" => "text", "regex" => ".*" ] ]
+                ]
+            ),
+            self::testpath() . 'valid/merge/merge_with_same_states_tokenizer_rules.json' => (
                 [
                     "start" => [ 0 => (object)[ "next" => "text-state" ] ],
                     "text-state" => [
@@ -579,13 +617,13 @@ class tokenizer_test extends \advanced_testcase {
 
     private static function setup_prepare_cases(): void {
         self::$preparetestcases = array(
-            self::testpath() . 'valid/prepare/prepare_with_one_state_highlight_rules.json' => (
+            self::testpath() . 'valid/prepare/prepare_with_one_state_tokenizer_rules.json' => (
                 [
                     "regexprs" => [ "start" => "/(\/\/)|(\/\*)|($)/" ],
                     "matchmappings" => [ "start" => [ "default_token" => "text", 0 => 0, 1 => 1 ] ]
                 ]
             ),
-            self::testpath() . 'valid/prepare/prepare_with_two_states_highlight_rules.json' => (
+            self::testpath() . 'valid/prepare/prepare_with_two_states_tokenizer_rules.json' => (
                 [
                     "regexprs" => [ "start" => "/(\/\/)|(\/\*)|($)/", "another_start" => "/(\/\/)|(\/\*)|($)/" ],
                     "matchmappings" => [
@@ -594,13 +632,13 @@ class tokenizer_test extends \advanced_testcase {
                     ]
                 ]
             ),
-            self::testpath() . 'valid/prepare/prepare_with_groups_highlight_rules.json' => (
+            self::testpath() . 'valid/prepare/prepare_with_groups_tokenizer_rules.json' => (
                 [
                     "regexprs" => [ "start" => "/(\/\/)|((?:.*)(?:b))|($)/" ],
                     "matchmappings" => [ "start" => [ "default_token" => "comment", 0 => 0, 1 => 1 ], ]
                 ]
             ),
-            self::testpath() . 'valid/prepare/prepare_with_more_rules_highlight_rules.json' => (
+            self::testpath() . 'valid/prepare/prepare_with_more_rules_tokenizer_rules.json' => (
                 [
                     "regexprs" => [
                         "start" => "/(\/\/)|((?:void)(?:[a-z]+(?:[a-zA-Z0-9]|_)*)(?:\()(?:\)))|($)/",
@@ -614,7 +652,7 @@ class tokenizer_test extends \advanced_testcase {
                     ]
                 ]
             ),
-            self::testpath() . 'valid/prepare/prepare_with_complex_matching_highlight_rules.json' => (
+            self::testpath() . 'valid/prepare/prepare_with_complex_matching_tokenizer_rules.json' => (
                 [
                     "regexprs" => [
                         "start" =>
@@ -630,19 +668,19 @@ class tokenizer_test extends \advanced_testcase {
                     ]
                 ]
             ),
-            self::testpath() . 'valid/prepare/prepare_with_one_group_highlight_rules.json' => (
+            self::testpath() . 'valid/prepare/prepare_with_one_group_tokenizer_rules.json' => (
                 [
                     "regexprs" => [ "start" => "/((?:\/\/))|($)/" ],
                     "matchmappings" => [ "start" => [ "default_token" => "text", 0 => 0 ] ]
                 ]
             ),
-            self::testpath() . 'valid/prepare/prepare_not_enough_groups_at_regex_highlight_rules.json' => (
+            self::testpath() . 'valid/prepare/prepare_not_enough_groups_at_regex_tokenizer_rules.json' => (
                 [
                     "regexprs" => [ "start" => "/((?:int))|($)/"],
                     "matchmappings" => [ "start" => [ "default_token" => "text", 0 => 0 ] ]
                 ]
             ),
-            self::testpath() . 'valid/prepare/prepare_with_number_ref_highlight_rules.json' => (
+            self::testpath() . 'valid/prepare/prepare_with_number_ref_tokenizer_rules.json' => (
                 [
                     "regexprs" => [ "start" => "/((a)(b)\\2\\3)|($)/" ],
                     "matchmappings" => [ "start" => [ "default_token" => "text", 0 => 0 ] ]
@@ -653,13 +691,13 @@ class tokenizer_test extends \advanced_testcase {
 
     private static function setup_preparse_cases(): void {
         self::$preparsetestcases = array(
-            self::testpath() . 'valid/get_all_tokens/no_line_highlight_rules.json' => (
+            self::testpath() . 'valid/get_all_tokens/no_line_tokenizer_rules.json' => (
                 [
                     'input' => self::testpath() . 'valid/get_all_tokens/no_line.c',
                     'output' => [ 0 => [ 'state' => 'start', 'tokens' => [] ] ]
                 ]
             ),
-            self::testpath() . 'valid/get_all_tokens/one_line_highlight_rules.json' => (
+            self::testpath() . 'valid/get_all_tokens/one_line_tokenizer_rules.json' => (
                 [
                     'input' => self::testpath() . 'valid/get_all_tokens/one_line.c',
                     'output' => [
@@ -670,7 +708,7 @@ class tokenizer_test extends \advanced_testcase {
                     ]
                 ]
             ),
-            self::testpath() . 'valid/get_all_tokens/two_lines_highlight_rules.json' => (
+            self::testpath() . 'valid/get_all_tokens/two_lines_tokenizer_rules.json' => (
                 [
                     'input' => self::testpath() . 'valid/get_all_tokens/two_lines.java',
                     'output' => [
@@ -688,7 +726,7 @@ class tokenizer_test extends \advanced_testcase {
                     ]
                 ]
             ),
-            self::testpath() . 'valid/get_all_tokens/more_lines_highlight_rules.json' => (
+            self::testpath() . 'valid/get_all_tokens/more_lines_tokenizer_rules.json' => (
                 [
                     'input' => self::testpath() . 'valid/get_all_tokens/more_lines.c',
                     'output' => [
@@ -804,7 +842,7 @@ class tokenizer_test extends \advanced_testcase {
 
     private static function setup_get_line_tokens_cases(): void {
         self::$getlinetokenstestcases = array(
-            self::testpath() . 'valid/get_line_tokens/no_matchs_highlight_rules.json' => (
+            self::testpath() . 'valid/get_line_tokens/no_matchs_tokenizer_rules.json' => (
                 [
                     'input' => '/* test comments',
                     'output' => [ 'state' => 'start', 'tokens' => array(
@@ -812,7 +850,7 @@ class tokenizer_test extends \advanced_testcase {
                     ) ]
                 ]
             ),
-            self::testpath() . 'valid/get_line_tokens/one_rule_highlight_rules.json' => (
+            self::testpath() . 'valid/get_line_tokens/one_rule_tokenizer_rules.json' => (
                 [
                     'input' => 'int',
                     'output' => [ 'state' => 'start', 'tokens' => array(
@@ -820,7 +858,7 @@ class tokenizer_test extends \advanced_testcase {
                     ) ]
                 ]
             ),
-            self::testpath() . 'valid/get_line_tokens/two_rules_highlight_rules.json' => (
+            self::testpath() . 'valid/get_line_tokens/two_rules_tokenizer_rules.json' => (
                 [
                     'input' => 'int ',
                     'output' => [ 'state' => 'start', 'tokens' => array(
@@ -828,7 +866,7 @@ class tokenizer_test extends \advanced_testcase {
                     ) ]
                 ]
             ),
-            self::testpath() . 'valid/get_line_tokens/more_rules_highlight_rules.json' => (
+            self::testpath() . 'valid/get_line_tokens/more_rules_tokenizer_rules.json' => (
                 [
                     'input' => 'int a = 10;',
                     'output' => [ 'state' => 'start', 'tokens' => array(
@@ -839,7 +877,7 @@ class tokenizer_test extends \advanced_testcase {
                     ) ]
                 ]
             ),
-            self::testpath() . 'valid/get_line_tokens/for_highlight_rules.json' => (
+            self::testpath() . 'valid/get_line_tokens/for_tokenizer_rules.json' => (
                 [
                     'input' => 'for (int i = 0; i < 10; i++) {',
                     'output' => [ 'state' => 'start', 'tokens' => array(
@@ -858,7 +896,7 @@ class tokenizer_test extends \advanced_testcase {
                     ) ]
                 ]
             ),
-            self::testpath() . 'valid/get_line_tokens/two_states_highlight_rules.json' => (
+            self::testpath() . 'valid/get_line_tokens/two_states_tokenizer_rules.json' => (
                 [
                     'input' => '/* test comments */',
                     'output' => [ 'state' => 'start', 'tokens' => array(
@@ -868,7 +906,7 @@ class tokenizer_test extends \advanced_testcase {
                     ) ]
                 ]
             ),
-            self::testpath() . 'valid/get_line_tokens/unexisted_state_highlight_rules.json' => (
+            self::testpath() . 'valid/get_line_tokens/unexisted_state_tokenizer_rules.json' => (
                 [
                     'input' => '// test comment',
                     'output' => [ 'state' => 'start', 'tokens' => array(
@@ -877,7 +915,7 @@ class tokenizer_test extends \advanced_testcase {
                     ) ]
                 ]
             ),
-            self::testpath() . 'valid/get_line_tokens/token_array_highlight_rules.json' => (
+            self::testpath() . 'valid/get_line_tokens/token_array_tokenizer_rules.json' => (
                 [
                     'input' => 'hello () {',
                     'output' => [ 'state' => 'start', 'tokens' => array(
@@ -890,7 +928,7 @@ class tokenizer_test extends \advanced_testcase {
                     ) ]
                 ]
             ),
-            self::testpath() . 'valid/get_line_tokens/token_array_two_rules_highlight_rules.json' => (
+            self::testpath() . 'valid/get_line_tokens/token_array_two_rules_tokenizer_rules.json' => (
                 [
                     'input' => 'hello () {',
                     'output' => [ 'state' => 'start', 'tokens' => array(
@@ -906,7 +944,7 @@ class tokenizer_test extends \advanced_testcase {
         );
 
         self::$getlinetokenoverflowstestcases = array(
-            self::testpath() . 'valid/get_line_tokens/no_matchs_highlight_rules.json' => (
+            self::testpath() . 'valid/get_line_tokens/no_matchs_tokenizer_rules.json' => (
                 [
                     'input' => [
                         'max_token_count' => 0,
@@ -917,7 +955,7 @@ class tokenizer_test extends \advanced_testcase {
                     ) ]
                 ]
             ),
-            self::testpath() . 'valid/get_line_tokens/one_rule_highlight_rules.json' => (
+            self::testpath() . 'valid/get_line_tokens/one_rule_tokenizer_rules.json' => (
                 [
                     'input' => [
                         'max_token_count' => 1,
@@ -934,25 +972,25 @@ class tokenizer_test extends \advanced_testcase {
 
     private static function setup_parse_cases(): void {
         self::$parsetestcases = array(
-            self::testpath() . 'valid/parse/no_line_highlight_rules.json' => (
+            self::testpath() . 'valid/parse/no_line_tokenizer_rules.json' => (
                 [
                     'input' => self::testpath() . 'valid/parse/no_line.c',
                     'output' => [ ]
                 ]
             ),
-            self::testpath() . 'valid/parse/one_line_highlight_rules.json' => (
+            self::testpath() . 'valid/parse/one_line_tokenizer_rules.json' => (
                 [
                     'input' => self::testpath() . 'valid/parse/one_line.c',
                     'output' => [ ]
                 ]
             ),
-            self::testpath() . 'valid/parse/two_lines_highlight_rules.json' => (
+            self::testpath() . 'valid/parse/two_lines_tokenizer_rules.json' => (
                 [
                     'input' => self::testpath() . 'valid/parse/two_lines.java',
                     'output' => [ ]
                 ]
             ),
-            self::testpath() . 'valid/parse/more_lines_highlight_rules.json' => (
+            self::testpath() . 'valid/parse/more_lines_tokenizer_rules.json' => (
                 [
                     'input' => self::testpath() . 'valid/parse/more_lines.c',
                     'output' => [